requirements.txt
asgiref==3.5.2
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
django-rest-swagger==2.2.0
django-stubs==1.13.1
django-stubs-ext==0.7.0
djangorestframework==3.14.0
djangorestframework-stubs==1.8.0
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
mysqlclient==2.1.1
openapi==1.1.0
openapi-codec==1.3.2
Pillow==9.3.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
requests==2.28.1
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tr0nz0d==0.1.4
types-Pillow==9.3.0.4
types-pytz==2022.7.0.0
types-PyYAML==6.0.12.2
types-requests==2.28.11.5
types-six==1.16.21.4
types-urllib3==1.26.25.4
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
